requirements.txt
elasticsearch==7
psycopg2
psycopg2.extras
loguru
pydantic
schedule
python-dotenv